requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
lazy-object-proxy==1.4.3
Pillow==8.0.1
PyJWT==1.7.1
